Redis 2.6 branch obtained from unstable removing all the cluster related code.
